Moved forecast api calls to repository. Refactored various dagger2 injection methods to separate modules. Various code cleanup.
