+++ app/src/androidTest/java/org/soaringforecast/rasp/AirportTest.java
@@ -34,7 +34,7 @@ public class AirportTest {
         appDatabase = Room.inMemoryDatabaseBuilder(context, AppDatabase.class)
                 .allowMainThreadQueries()
                 .build();
-        appRepository = AppRepository.getAppRepository(context);
+        appRepository = AppRepository.getAppRepository(context, null, null, null, null, null);
 
     }
