+++ app/src/main/java/org/soaringforecast/rasp/dagger/StringsUtilsModule.java
@@ -1,0 +1,20 @@
+package org.soaringforecast.rasp.dagger;
+
+import org.soaringforecast.rasp.utils.StringUtils;
+
+import javax.inject.Singleton;
+
+import dagger.Module;
+import dagger.Provides;
+
+@Module
+public class StringsUtilsModule {
+
+    @Provides
+    @Singleton
+    public StringUtils provideStringUtils() {
+        return new StringUtils();
+    }
+
+
+}
